Remove PatternEditorCubit
