+++ lib/widgets/editors/pattern_editor/pattern_editor.dart
@@ -19,18 +19,20 @@
 
 import 'dart:math';
 
+import 'package:anthem/model/project.dart';
+import 'package:anthem/theme.dart';
 import 'package:anthem/widgets/basic/background.dart';
 import 'package:anthem/widgets/basic/button.dart';
 import 'package:anthem/widgets/basic/dropdown.dart';
+import 'package:anthem/widgets/basic/icon.dart';
 import 'package:anthem/widgets/basic/menu/menu.dart';
 import 'package:anthem/widgets/basic/menu/menu_model.dart';
 import 'package:anthem/widgets/basic/scroll/scrollbar.dart';
-import 'package:anthem/widgets/editors/pattern_editor/pattern_editor_cubit.dart';
+import 'package:anthem/widgets/editors/pattern_editor/pattern_editor_controller.dart';
 import 'package:flutter/widgets.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:flutter_mobx/flutter_mobx.dart';
+import 'package:provider/provider.dart';
 
-import '../../../theme.dart';
-import '../../basic/icon.dart';
 import 'generator_row.dart';
 
 class PatternEditor extends StatefulWidget {
@@ -44,6 +46,8 @@ class _PatternEditorState extends State<PatternEditor> {
   double nextHue = 0;
   ScrollController verticalScrollController = ScrollController();
 
+  PatternEditorController? controller;
+
   Color getColor() {
     final color = HSLColor.fromAHSL(1, nextHue, 0.33, 0.5).toColor();
     nextHue = (nextHue + 330) % 360;
@@ -52,11 +56,14 @@ class _PatternEditorState extends State<PatternEditor> {
 
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<PatternEditorCubit, PatternEditorState>(
-        builder: (context, state) {
-      final menuController = MenuController();
+    final project = Provider.of<ProjectModel>(context);
+    controller ??= PatternEditorController(project: project);
 
-      return NotificationListener<SizeChangedLayoutNotification>(
+    final menuController = MenuController();
+
+    return Provider.value(
+      value: controller!,
+      child: NotificationListener<SizeChangedLayoutNotification>(
         onNotification: (notification) {
           WidgetsBinding.instance.addPostFrameCallback((_) {
             verticalScrollController.position.notifyListeners();
@@ -84,9 +91,7 @@ class _PatternEditorState extends State<PatternEditor> {
                             AnthemMenuItem(
                               text: "New pattern",
                               onSelected: () {
-                                final cubit =
-                                    context.read<PatternEditorCubit>();
-                                cubit.addPattern();
+                                controller!.addPattern();
                               },
                             )
                           ],
@@ -101,24 +106,25 @@ class _PatternEditorState extends State<PatternEditor> {
                         ),
                       ),
                       const SizedBox(width: 4),
-                      Dropdown(
-                        width: 169,
-                        height: 26,
-                        items: state.patternList.map(
-                          (item) {
-                            return DropdownItem(
-                              id: item.id.toString(),
-                              name: item.name,
-                            );
+                      Observer(builder: (context) {
+                        return Dropdown(
+                          width: 169,
+                          height: 26,
+                          items: project.song.patternOrder.map(
+                            (id) {
+                              final pattern = project.song.patterns[id]!;
+                              return DropdownItem(
+                                id: id,
+                                name: pattern.name,
+                              );
+                            },
+                          ).toList(),
+                          selectedID: project.song.activePatternID?.toString(),
+                          onChanged: (id) {
+                            project.song.activePatternID = id;
                           },
-                        ).toList(),
-                        selectedID: state.activePatternID?.toString(),
-                        onChanged: (id) {
-                          context
-                              .read<PatternEditorCubit>()
-                              .setActivePattern(id);
-                        },
-                      ),
+                        );
+                      }),
                       const Expanded(child: SizedBox()),
                     ],
                   ),
@@ -139,27 +145,18 @@ class _PatternEditorState extends State<PatternEditor> {
                                 padding:
                                     const EdgeInsets.symmetric(vertical: 5),
                                 child: SizeChangedLayoutNotifier(
-                                  child: Column(
-                                    children:
-                                        state.generatorIDList.map<Widget>((id) {
-                                      final instrument = state.instruments[id];
-                                      final controller = state.controllers[id];
-
-                                      if (instrument != null) {
-                                        return GeneratorRow(generatorID: id);
-                                      }
-
-                                      if (controller != null) {
+                                  child: Observer(builder: (context) {
+                                    return Column(
+                                      children: project.generatorList
+                                          .map<Widget>((id) {
                                         return Padding(
                                           padding:
                                               const EdgeInsets.only(bottom: 1),
                                           child: GeneratorRow(generatorID: id),
                                         );
-                                      }
-
-                                      throw Error();
-                                    }).toList(),
-                                  ),
+                                      }).toList(),
+                                    );
+                                  }),
                                 ),
                               ),
                             ),
@@ -186,10 +183,10 @@ class _PatternEditorState extends State<PatternEditor> {
                           contentPadding: EdgeInsets.zero,
                           startIcon: Icons.add,
                           onPress: () {
-                            context.read<PatternEditorCubit>().addGenerator(
-                                  "Instrument ${(Random()).nextInt(100).toString()}",
-                                  getColor(),
-                                );
+                            controller!.addGenerator(
+                              "Instrument ${(Random()).nextInt(100).toString()}",
+                              getColor(),
+                            );
                           },
                         ),
                         const SizedBox(width: 24),
@@ -202,7 +199,7 @@ class _PatternEditorState extends State<PatternEditor> {
             ),
           ),
         ),
-      );
-    });
+      ),
+    );
   }
 }
